replace float16 with bfloat16

+++ eagle/evaluation/gen_baseline_answer_mix.py
@@ -161,7 +161,7 @@ def get_model_answers(
         base_model_path=base_model_path,
         ea_model_path=ea_model_path,
         Type="Mixtral",
-        torch_dtype=torch.float16,
+        torch_dtype=torch.bfloat16,
         low_cpu_mem_usage=True,
         # load_in_8bit=True,
         device_map="auto"
